+++ tests/run_tests.sh
@@ -197,7 +197,7 @@ do
       || LAST_EXIT_CODE=$?
       ;;
     "update")
-         $APPLICATION_ROOT/.github/actions/init_initialize-0.80-db.sh \
+         $APPLICATION_ROOT/.github/actions/init_initialize-0.85.5-db.sh \
       && $APPLICATION_ROOT/.github/actions/init_initialize-9.5-db.sh \
       && docker-compose exec -T app .github/actions/test_update-from-older-version.sh \
       && docker-compose exec -T app .github/actions/test_update-from-9.5.sh \